OpenID4VP: Move out AuthnResponseResult into separate file
